@@ -6,22 +6,22 @@ homepage: https://leadcode.dev
 repository: https://github.com/bratan/flutter_translate_gen
 
 environment:
-  sdk: ">=2.8.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   dart_casing: ^1.0.1
   dart_utils: ^1.0.2
-  source_gen: '>=0.9.4 <1.0.0'
-  build: '>=1.2.1 <2.0.0'
+  source_gen: ^1.0.0
+  build: ^2.0.0
   code_builder: '>=3.2.0 <4.0.0'
-  analyzer: '>=0.36.4 <1.0.0'
-  glob: ^1.2.0
+  analyzer: ^1.4.0
+  glob: ^2.0.1
   # TODO point to 1.0.2 as soon as it is merged
   flutter_translate_annotations:
     git: https://github.com/avbk/flutter_translate_annotations
 
 
 dev_dependencies:
-  build_runner: '>=1.7.0 <2.0.0'
-  lint: ^1.2.0
-  test: ^1.15.2
+  build_runner: ^1.12.2
+  lint: ^1.5.3
+  test: ^1.16.8
