@@ -1,4 +1,5 @@
-﻿import 'dart:async';
+﻿// @dart=2.9
+import 'dart:async';
 
 import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
